Add files via upload
Don't worry about TriggerEnter, TriggerExit or WallEnter. These are codes i made for the big coral wall that lies outside the coral reef, TEn and TEx are children of WEn, so theyre all meant to be used together. 
Also for now you may delete anything in the Testmove script that includes Text, TMPro or SurfaceLevel if it hinders the main code from functioning properly.